chore: split deploy-camunda matrix runner.go into concern-scoped files
`scripts/deploy-camunda/matrix/runner.go` was 2705 lines and mixed deploy
execution, dispatch, hook orchestration, namespace lifecycle, diagnostics,
dry-run/coverage rendering, OCI immutability handling, and result reporting.
Epic #6264 flagged it as the second structural merge-conflict surface
(9 of 100 sampled commits).

Split the three actual churn drivers into their own files inside the
existing `package matrix`:

- `runner_options.go` (160 lines) — `RunOptions` struct, grows linearly
  with every new CLI flag.
- `runner_execute.go` (613 lines) — `executeEntry`, the orchestration
  glue every new feature touches. Entra/Auth0 provisioning blocks left
  inlined unchanged.
- `runner_upgrade.go` (404 lines) — `executeTwoStepUpgrade`,
  `executeUpgradeOnly`, bitnami PG password helpers.

`runner.go` shrinks 2705 → 1560 lines. Pure move: no signature changes,
no helper introductions, no test updates. Same package, zero import
churn for `cmd/matrix.go` or any existing test file.

Verification:
- `go build ./...` green.
- `go test ./matrix/...` green.
- `make go.test chartPath=charts/camunda-platform-8.10` green.
- `deploy-camunda matrix list --format json` diff pre/post: empty.
- `deploy-camunda matrix run --dry-run` diff pre/post: timestamp only.
- `deploy-camunda matrix run --coverage` diff pre/post: timestamp only.

Refs: #6303, ADR 0093 ("Follow-ups").
